Fix some errors due to refactoring
